start

+++ src/main/java/com/github/andriell/db/Product.java
@@ -63,7 +63,7 @@ public class Product {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(float price) {
         this.price = price;
     }
